Refactor quotas logic into container
